+++ lib/validation.dart
@@ -57,6 +57,15 @@ class UuidValidation {
           final match = regex.hasMatch(fromString.toLowerCase());
           return match;
         }
+      case ValidationMode.loose:
+        {
+          var pattern = (noDashes)
+              ? r'^([0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}|[0-9a-f]{32})$'
+              : r'^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$';
+          final regex = RegExp(pattern, caseSensitive: false, multiLine: true);
+          final match = regex.hasMatch(fromString.toLowerCase());
+          return match;
+        }
     }
   }
 
@@ -65,7 +74,8 @@ class UuidValidation {
   /// You can choose to validate from a string or from a byte list based on
   /// which parameter is passed.
   /// Optionally you can set [validationMode] to `ValidationMode.nonStrict` to
-  /// allow for non RFC4122 compliant UUIDs.
+  /// allow for non RFC4122 compliant UUIDs, or to `ValidationMode.loose` to
+  /// allow any 128 bit uuid-like id (e.g. ulid or future versions of uuid).
   /// If you are using a Microsoft GUID, you should set [validationMode] to
   /// `ValidationMode.nonStrict`.
   static void isValidOrThrow(
@@ -91,6 +101,18 @@ class UuidValidation {
         if (isValidNonStrict) {
           throw FormatException(
               'The provided UUID is not RFC4122 compliant. It seems you might be using a Microsoft GUID. Try setting `validationMode = ValidationMode.nonStrict`',
+              fromString);
+        }
+
+        final isValidLoose = isValidUUID(
+            fromString: fromString,
+            fromByteList: fromByteList,
+            validationMode: ValidationMode.loose,
+            noDashes: noDashes);
+
+        if (isValidLoose) {
+          throw FormatException(
+              'The provided UUID is not RFC4122 compliant. Try setting `validationMode = ValidationMode.loose`',
               fromString);
         }
       }