+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+Unreleased
+* Add support for loose validation
+
 v4.5.3
 
 * Remove experimental flag on UUIDValue and remote meta dependency. (thanks @marcelomendoncasoares)